Frames.Col

Description

Column types

Documentation

type (:->) (a :: Symbol) b = '(a, b)

pattern Col :: KnownSymbol s => t -> ElField '(s, t)

getCol :: ElField '(s, t) -> t

Get the data payload of a named field.

newtype Col' s a

Used only for a show instance that parenthesizes the value.

Constructors

Col' (ElField (s :-> a))

col' :: KnownSymbol s => a -> Col' s a

Helper for making a Col'

type family ReplaceColumns x ys where ...

ReplaceColumns x ys keeps the textual name of each element of ys, but replaces its data type with x.

Equations

ReplaceColumns x '[] = '[]

ReplaceColumns x ((c :-> y) ': ys) = (c :-> x) ': ReplaceColumns x ys