Frames.ColumnTypeable

Documentation

data Parsed a

Constructors

Possibly a

Definitely a

parsedValue :: Parsed a -> a

class Parseable a where

Values that can be read from a Text with more or less discrimination.

Minimal complete definition

Nothing

Methods

parse :: MonadPlus m => Text -> m (Parsed a)

Returns Nothing if a value of the given type can not be read; returns 'Just Possibly' if a value can be read, but is likely ambiguous (e.g. an empty string); returns 'Just Definitely' if a value can be read and is unlikely to be ambiguous."

default parse :: (Readable a, MonadPlus m) => Text -> m (Parsed a)

parseCombine :: MonadPlus m => Parsed a -> Parsed a -> m (Parsed a)

Combine two parse results such that the combination can fail. Useful when we have two Possibly parsed values that are different enough to suggest the parse of each should be considered a failure. The default implementation is to return the first argument.

default parseCombine :: MonadPlus m => Parsed a -> Parsed a -> m (Parsed a)

representableAsType :: Parsed a -> Const (Either (String -> Q [Dec]) Type) a

default representableAsType :: Typeable a => Parsed a -> Const (Either (String -> Q [Dec]) Type) a

discardConfidence :: Parsed a -> a

Discard any estimate of a parse's ambiguity.

parse' :: (MonadPlus m, Parseable a) => Text -> m a

Acts just like fromText: tries to parse a value from a Text and discards any estimate of the parse's ambiguity.

parseIntish :: (Readable a, MonadPlus f) => Text -> f (Parsed a)

class ColumnTypeable a where

This class relates a universe of possible column types to Haskell types, and provides a mechanism to infer which type best represents some textual data.

Methods

colType :: a -> Either (String -> Q [Dec]) Type

inferType :: Text -> a