Frames.ColumnUniverse

Documentation

data CoRec (a :: k -> Type) (b :: [k])

Generalize algebraic sum types.

type Columns = ColumnUniverse CommonColumns

A universe of common column variants. These are the default column types that Frames can infer. See the Tutorial for an example of extending the default types with your own.

type ColumnUniverse = CoRec ColInfo

Define a set of variants that captures all possible column types.

data ColInfo a

Information necessary for synthesizing row types and comparing types.

type CommonColumns = [Bool, Int, Double, Text]

Common column types: Bool, Int, Double, Text

type CommonColumnsCat = [Bool, Int, Double, Categorical 8, Text]

Common column types including categorical types.

parsedTypeRep :: ColInfo a -> Parsed Type