Frames.Frame

Description

A Frame is a finite Int-indexed collection of rows.

Documentation

data Frame r

A Frame is a finite collection of rows indexed by Int.

Constructors

Frame

Fields

frameLength :: !Int

frameRow :: Int -> r

type FrameRec rs = Frame (Record rs)

A Frame whose rows are Record values.

boxedFrame :: Foldable f => f r -> Frame r

Build a Frame from any Foldable. This simply uses a boxed Vector to hold each row. If you have a collection of Records, consider using toFrame.

zipFrames :: FrameRec rs -> FrameRec rs' -> FrameRec (rs ++ rs')

Horizontal Frame concatenation. That is, zipFrames f1 f2 will return a Frame with as many rows as the smaller of f1 and f2 whose rows are the result of appending the columns of f2 to those of f1.