Frames.RecF

Documentation

rappend :: forall k (f :: k -> Type) (as :: [k]) (bs :: [k]). Rec f as -> Rec f bs -> Rec f (as ++ bs)

Two records may be pasted together.

rtraverse :: forall u h f g (rs :: [u]). Applicative h => (forall (x :: u). f x -> h (g x)) -> Rec f rs -> h (Rec g rs)

A record may be traversed with respect to its interpretation functor. This can be used to yank (some or all) effects from the fields of the record to the outside of the record.

rdel :: CanDelete r rs => Rec f rs -> Rec f (RDelete r rs)

Delete a field from a record

type CanDelete r rs = (RElem r rs (RIndex r rs), RDelete r rs ⊆ rs)

A constraint that a field can be deleted from a record.

type AllCols c ts = AllConstrained c (UnColumn ts)

Enforce a constraint on the payload type of each column.

type family UnColumn (ts :: [(Symbol, *)]) where ...

Strip the column information from each element of a list of types.

Equations

UnColumn '[] = '[]

UnColumn ((s :-> t) ': ts) = t ': UnColumn ts

class StripFieldNames (ts :: [(Symbol, Type)]) where

Facilities for removing and replacing the type-level label, or column name, part of a record.

Methods

stripNames :: Rec ElField ts -> Rec Identity (Unlabeled ts)

stripNames' :: forall (f :: Type -> Type). Functor f => Rec (f :. ElField) ts -> Rec f (Unlabeled ts)

withNames :: Rec Identity (Unlabeled ts) -> Rec ElField ts

withNames' :: forall (f :: Type -> Type). Functor f => Rec f (Unlabeled ts) -> Rec (f :. ElField) ts

mapMono :: (AllAre a (Unlabeled ts), StripFieldNames ts, StripFieldNames (ReplaceAllSnd b ts), ReplaceAll b (Unlabeled ts) ~ Unlabeled (ReplaceAllSnd b ts)) => (a -> b) -> Rec ElField ts -> Rec ElField (ReplaceAllSnd b ts)

Map a function across a homogeneous Rec of named values. This is a thin wrapper over mapMonoV.

type family ColFun f x where ...

A type function to convert a 'Rec f' to a 'Rec (g :. f)'. ColFun f (Record rs) = Rec (f :. ElField) rs.

Equations

ColFun f (Rec g rs) = Rec (f :. g) rs

class ColumnHeaders (cs :: [(Symbol, *)])

Minimal complete definition

columnHeaders

columnHeaders :: ColumnHeaders cs => proxy (Rec f cs) -> [String]

Return the column names for a record.